+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Release Notes - heidelpay Payment Gateway for Magento 2
 
-## v17.4.X
+## v17.4.13
 
 ### Added
 - Invoice and Secured Invoice (B2C) Payment Methods (DE, AT, CH)
+- Sending Finalize notifications to heidelpay when Shipment is created (Invoice Methods only)
+- Added Push functionality
 
 ### Changed
 - Invoices now contain information about payments (e.g. where to send the amount, in case of Prepayment, Invoice)
+- When errors after sending the request or during the redirect are occuring, the customer will be redirected to the checkout/cart instead of seeing a blank page
 
 ## v17.3.28